#95 refactored namespaces and made certain properties and fields nullable where necessary.
