+++ backend/WebApi/Api/Model/FriendConnection.cs
@@ -1,4 +1,6 @@
-﻿namespace WebApi.Model;
+﻿using WebApi.Model;
+
+namespace WebApi.Api.Model;
 
 public class FriendConnection
 {
