+++ backend/WebApi/Api/Model/User.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using WebApi.Api.Model;
 
 namespace WebApi.Model;
