+++ backend/WebApi/Api/Service/AuthService.cs
@@ -140,6 +140,6 @@ public class AuthService : IAuthService
 public class AuthResult
 {
     public bool Success { get; set; }
-    public string Token { get; set; }
-    public string ErrorMessage { get; set; }
+    public string? Token { get; set; }
+    public string? ErrorMessage { get; set; }
 }
