+++ backend/WebApi/Api/Service/UserService.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Mvc;
 using WebApi.Api.Common;
+using WebApi.Api.Model;
 using WebApi.Model;
 
 namespace WebApi;
